fix code and template to edit tags

+++ psunote/noteapp.py
@@ -19,10 +19,7 @@ def index():
     notes = db.session.execute(
         db.select(models.Note).order_by(models.Note.title)
     ).scalars()
-    return flask.render_template(
-        "index.html",
-        notes=notes,
-    )
+    return flask.render_template("index.html", notes=notes)
 
 
 # Create note
@@ -54,10 +51,7 @@ def notes_create():
         db.session.commit()
         return flask.redirect(flask.url_for("index"))
 
-    return flask.render_template(
-        "notes-create.html",
-        form=form,
-    )
+    return flask.render_template("notes-create.html", form=form)
 
 
 # create tag
@@ -88,33 +82,16 @@ def notes_edit(note_id):
     form = forms.NoteForm(obj=note)
 
     if form.validate_on_submit():
-        form.populate_obj(note)
-        note.tags = []
-
-        tag_names = (
-            [name.strip() for name in form.tags.data.split(",")]
-            if form.tags.data
-            else []
-        )
-        for tag_name in tag_names:
-            tag = db.session.execute(
-                db.select(models.Tag).where(models.Tag.name == tag_name)
-            ).scalar_one_or_none()
-
-            if not tag:
-                tag = models.Tag(name=tag_name)
-                db.session.add(tag)
-
-            note.tags.append(tag)
+        # Update only the fields that are not empty
+        if form.title.data:
+            note.title = form.title.data
+        if form.description.data:
+            note.description = form.description.data
 
         db.session.commit()
         return flask.redirect(flask.url_for("index"))
 
-    return flask.render_template(
-        "notes-edit.html",
-        form=form,
-        note=note,
-    )
+    return flask.render_template("notes-edit.html", form=form, note=note)
 
 
 # delete note
@@ -138,15 +115,12 @@ def tags_edit(tag_name):
     form = forms.TagForm(obj=tag)
 
     if form.validate_on_submit():
-        form.populate_obj(tag)
-        db.session.commit()
-        return flask.redirect(flask.url_for("tags_view", tag_name=tag.name))
+        if form.name.data and form.name.data != tag.name:
+            tag.name = form.name.data
+            db.session.commit()
+            return flask.redirect(flask.url_for("tags_view", tag_name=tag.name))
 
-    return flask.render_template(
-        "tags-edit.html",
-        form=form,
-        tag=tag,
-    )
+    return flask.render_template("tags-edit.html", form=form, tag=tag)
 
 
 # delete tag
@@ -157,8 +131,10 @@ def tags_delete(tag_name):
         db.select(models.Tag).where(models.Tag.name == tag_name)
     ).scalar_one_or_none()
 
-    db.session.delete(tag)
-    db.session.commit()
+    if tag:
+        db.session.delete(tag)
+        db.session.commit()
+
     return flask.redirect(flask.url_for("index"))
